move shared object into own module
