+++ pygfx/renderers/wgpu/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 
 from ._utils import registry, register_wgpu_render_function
-from ._renderer import stdinfo_uniform_type
+from ._shared import stdinfo_uniform_type
 from ._renderer import WgpuRenderer
 from . import meshshader
